Color grid cell on mouse over

+++ Horde/Assets/Scripts/MapController.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 
 public class MapController : MonoBehaviour {
 	public float cellSize = 32;
-	public float xCell = 10f;
-	public float yCell = 10f;
+	public int xCell = 10;
+	public int yCell = 10;
 
 	// Use this for initialization
 	void Start () {
-		transform.localScale = new Vector3(xCell, yCell, 1);
+		transform.localScale = new Vector3((float) xCell, (float) yCell, 1);
 	}
 	
 	// Update is called once per frame
